Sagas and actions rework
