+++ src/actions/experimentAC.js
@@ -1,3 +1,9 @@
+/**
+ * These are Action Creators for Experiments state.
+ * Both dispatchin data from components to Sagas and passing data from Sagas
+ * to experiment Reducer.
+ */
+ 
 export function getExperiments(id) {
   return {
     type: 'GET_EXPERIMENTS',
